Use Configuration object from eyes.webdriver
